+++ com.faolline.graphdialoguesystem/Tests/EditMode/Runtime/ActionTests.cs
@@ -2,6 +2,7 @@ using NUnit.Framework;
 using UnityEngine;
 using Faolline.GraphCore;
 using Faolline.GraphDialogue;
+using SetBoolAction = Faolline.GraphDialogue.SetBoolAction;   // disambiguate from Faolline.GraphCore.SetBoolAction
 
 namespace Faolline.GraphDialogue.Tests
 {