+++ com.faolline.graphdialoguesystem/Tests/EditMode/Runtime/InlineConditionPlaybackTests.cs
@@ -3,6 +3,7 @@ using NUnit.Framework;
 using UnityEngine;
 using Faolline.GraphCore;
 using Faolline.GraphDialogue;
+using BoolCondition = Faolline.GraphDialogue.BoolCondition;   // disambiguate from Faolline.GraphCore.BoolCondition
 
 namespace Faolline.GraphDialogue.Tests
 {